+++ bcmc/src/main/java/com/adyen/checkout/bcmc/BcmcComponent.kt
@@ -27,6 +27,7 @@ import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.cse.CardEncrypter
 import com.adyen.checkout.cse.UnencryptedCard
 import com.adyen.checkout.cse.exception.EncryptionException
+import com.adyen.threeds2.ThreeDS2Service
 import kotlinx.coroutines.launch
 
 private val TAG = LogUtil.getTag()
@@ -78,7 +79,8 @@ class BcmcComponent(
         Logger.v(TAG, "onInputDataChanged")
         return BcmcOutputData(
             validateCardNumber(inputData.cardNumber),
-            validateExpiryDate(inputData.expiryDate)
+            validateExpiryDate(inputData.expiryDate),
+            inputData.isStorePaymentSelected
         )
     }
 
@@ -121,8 +123,15 @@ class BcmcComponent(
             encryptedCardNumber = encryptedCard.encryptedCardNumber
             encryptedExpiryMonth = encryptedCard.encryptedExpiryMonth
             encryptedExpiryYear = encryptedCard.encryptedExpiryYear
+            try {
+                threeDS2SdkVersion = ThreeDS2Service.INSTANCE.sdkVersion
+            } catch (e: ClassNotFoundException) {
+                Logger.e(TAG, "threeDS2SdkVersion not set because 3DS2 SDK is not present in project.")
+            }
         }
         paymentComponentData.paymentMethod = cardPaymentMethod
+        paymentComponentData.setStorePaymentMethod(outputData.isStoredPaymentMethodEnabled)
+        paymentComponentData.shopperReference = configuration.shopperReference
         return GenericComponentState(paymentComponentData, true, true)
     }