+++ components-core/src/main/java/com/adyen/checkout/components/base/BaseActionComponent.java
@@ -66,8 +66,28 @@ public abstract class BaseActionComponent<ConfigurationT extends Configuration> 
     }
 
     @Override
+    public void removeObservers(@NonNull LifecycleOwner lifecycleOwner) {
+        mResultLiveData.removeObservers(lifecycleOwner);
+    }
+
+    @Override
+    public void removeObserver(@NonNull final Observer<ActionComponentData> observer) {
+        mResultLiveData.removeObserver(observer);
+    }
+
+    @Override
     public void observeErrors(@NonNull LifecycleOwner lifecycleOwner, @NonNull Observer<ComponentError> observer) {
         mErrorMutableLiveData.observe(lifecycleOwner, observer);
+    }
+
+    @Override
+    public void removeErrorObservers(@NonNull LifecycleOwner lifecycleOwner) {
+        mErrorMutableLiveData.removeObservers(lifecycleOwner);
+    }
+
+    @Override
+    public void removeErrorObserver(@NonNull final Observer<ComponentError> observer) {
+        mErrorMutableLiveData.removeObserver(observer);
     }
 
     /**