+++ components-core/src/main/java/com/adyen/checkout/components/base/BasePaymentComponent.java
@@ -76,8 +76,28 @@ public abstract class BasePaymentComponent<
     }
 
     @Override
+    public void removeObservers(@NonNull LifecycleOwner lifecycleOwner) {
+        mPaymentComponentStateLiveData.removeObservers(lifecycleOwner);
+    }
+
+    @Override
+    public void removeObserver(@NonNull final Observer<ComponentStateT> observer) {
+        mPaymentComponentStateLiveData.removeObserver(observer);
+    }
+
+    @Override
     public void observeErrors(@NonNull LifecycleOwner lifecycleOwner, @NonNull Observer<ComponentError> observer) {
         mComponentErrorLiveData.observe(lifecycleOwner, observer);
+    }
+
+    @Override
+    public void removeErrorObservers(@NonNull LifecycleOwner lifecycleOwner) {
+        mComponentErrorLiveData.removeObservers(lifecycleOwner);
+    }
+
+    @Override
+    public void removeErrorObserver(@NonNull final Observer<ComponentError> observer) {
+        mComponentErrorLiveData.removeObserver(observer);
     }
 
     @Override