+++ googlepay/src/main/java/com/adyen/checkout/googlepay/model/GooglePayParamUtils.kt
@@ -1,0 +1,14 @@
+package com.adyen.checkout.googlepay.model
+
+import com.adyen.checkout.googlepay.util.AllowedCardNetworks
+
+internal object GooglePayParamUtils {
+
+    fun mapBrandToGooglePayNetwork(brand: String): String? {
+        return when {
+            brand == "mc" -> AllowedCardNetworks.MASTERCARD
+            AllowedCardNetworks.getAllAllowedCardNetworks().contains(brand.uppercase()) -> brand.uppercase()
+            else -> null
+        }
+    }
+}