+++ googlepay/src/main/java/com/adyen/checkout/googlepay/model/GooglePayParams.kt
@@ -10,14 +10,20 @@ package com.adyen.checkout.googlepay.model
 
 import com.adyen.checkout.components.model.payments.Amount
 import com.adyen.checkout.core.exception.ComponentException
+import com.adyen.checkout.core.log.LogUtil
+import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.googlepay.GooglePayConfiguration
+import com.adyen.checkout.googlepay.util.AllowedCardNetworks
+
+private val TAG = LogUtil.getTag()
 
 /**
  * Model class holding the parameters required to build requests for GooglePay
  */
 data class GooglePayParams(
     private val googlePayConfiguration: GooglePayConfiguration,
-    private val serverGatewayMerchantId: String?
+    private val serverGatewayMerchantId: String?,
+    private val availableCardNetworksFromApi: List<String>?
 ) {
     val gatewayMerchantId: String = getPreferredGatewayMerchantId()
     val googlePayEnvironment: Int = googlePayConfiguration.googlePayEnvironment
@@ -26,7 +32,7 @@ data class GooglePayParams(
     val countryCode: String? = googlePayConfiguration.countryCode
     val merchantInfo: MerchantInfo? = googlePayConfiguration.merchantInfo
     val allowedAuthMethods: List<String>? = googlePayConfiguration.allowedAuthMethods
-    val allowedCardNetworks: List<String>? = googlePayConfiguration.allowedCardNetworks
+    val allowedCardNetworks: List<String> = getAvailableCardNetworks()
     val isAllowPrepaidCards: Boolean = googlePayConfiguration.isAllowPrepaidCards
     val isEmailRequired: Boolean = googlePayConfiguration.isEmailRequired
     val isExistingPaymentMethodRequired: Boolean = googlePayConfiguration.isExistingPaymentMethodRequired
@@ -44,5 +50,20 @@ data class GooglePayParams(
             ?: throw ComponentException(
                 "GooglePay merchantAccount not found. Update your API version or pass it manually inside your GooglePayConfiguration"
             )
+    }
+
+    private fun getAvailableCardNetworks(): List<String> {
+        return googlePayConfiguration.allowedCardNetworks
+            ?: getAvailableCardNetworksFromApi()
+            ?: AllowedCardNetworks.getAllAllowedCardNetworks()
+    }
+
+    private fun getAvailableCardNetworksFromApi(): List<String>? {
+        if (availableCardNetworksFromApi == null) return null
+        return availableCardNetworksFromApi.mapNotNull { brand ->
+            val network = GooglePayParamUtils.mapBrandToGooglePayNetwork(brand)
+            if (network == null) Logger.e(TAG, "skipping brand $brand, as it is not an allowed card network.")
+            return@mapNotNull network
+        }
     }
 }