+++ RELEASE_NOTES.md
@@ -8,5 +8,16 @@
 [//]: # ( # Deprecated)
 [//]: # ( - Configurations public constructor are deprecated, please use each Configuration's builder to make a Configuration object)
 
+## Added
+- Dual-branded card flow: shoppers can now select their card's brand when two brands are detected.
+- Ability to store BCMC cards.
+- Support for new Activity Result API.
+- Ability to stop observing components.
+- 3DS2 SDK version in `CardComponent`'s output.
+- Read list of supported card networks for Google Pay from the payment methods API response.
+
 ## Fixed
-- Address Visibility in Card Component configuration not being reflected on UI.
+- Crash caused by having stored payment methods none of which is Ecommerce.
+- Google Pay Component will not include `TotalPrice` in its output, if `TotalPriceStatus` is set to `NOT_CURRENTLY_KNOWN`.
+- Issue in Drop-in when multiple payment methods have the same type.
+- Missing default BCMC configuration in Drop-in.